restructure + add a unit test on building message
